Use the API to check for login verification status
This makes the UI for the user stable, as unnecessary reloads are not
needed anymore. The UI requests the current status of the login
verification on the device using a new API instead of simply refreshing
the page every 3 seconds, which, depending on the connection speed, may
result in rendering issues and stuff like that.

However, this feature is experimental for now and needs to be enabled by
a feature flag explicitly. This is because API connections are hold open
for up to 30 seconds, until an actual result may be returned (as long as
the login verification request is not confirmed on the device earlier),
in order to save HTTP round-trips. This may result in some more
resources blocked on the server-side, as the connection need to be hold
open for this time.

Change-Id: I1f6a59e6fe2872b34152c72c116ca7d2d29d4741